validator: Revert --wait-for-exit flag on exit subcommand (#6723)
As written, agave-validator exit will always error out when run against
a running agave-validator bin that did not have the change. The exit is
actually initiated, but this could break automation that depends on the
return code of agave-validator exit.

The flag will be re-added to be more upgrade friendly in the future