accounts-db: Remove usage of RentCollector (#6781)
#### Problem

The RentCollector type is included in accounts-db types, but it's not
actually used for anything.

The only place it *is* used is to figure out the number of slots in an
epoch, but it looks like that call site only ever passes in `0` anyway,
from `RentCollector::default()`.

#### Summary of changes

Remove RentCollector from types and interfaces in accounts-db.