Make chained merkle required in Shredder (#8079)
The chained merkle root hash argument to Shredder is currently an
Option<Hash>. However, chained merkle shreds are active on all clusters
so this is always a Some(_) in production.

To continue the cleanup of shreds, remove the Option<_> from the
argument at Shredder and update all callers. Some underlying shred code
still takes an Option<Hash>, but that will be addressed at a later time