fetch-core-bpf: Add "v" to release name (#8195)
#### Problem

The core bpf download script doesn't work for all programs, because some
include a "v" in the release name, and others don't. For example, we
have `program@v1.0.0` for the stake program, and `program@3.0.0` for the
address lookup table program. This makes genesis fail.

#### Summary of changes

I created dual releases / tags for each of the four core bpf programs,
with a "v" and without a "v".

Since using the "v" is more common across the program repos, change the
fetch script to use that one, meaning: fetch `program@v1.0.0` instead of
`program@1.0.0`.